deformer/pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "wheel"
]
build-backend = "setuptools.build_meta"


[project]
name = "deformer"
version = "1.4.2"
description = "DEFormer - Development Environment Formatter"
readme = "README.md"
authors = [
    { name = "Tech Team" }
]
license = { text = "" }
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "pyyaml==5.3.1",
    "python-logstash-async==1.6.4",
    "tqdm==4.65.0",
    "requests>=2.28.0,<2.29"
]
requires-python = "~=3.8"


[project.scripts]
deformer = "deformer.deformer:main"


